+++ src/main/java/id/ac/ui/cs/advprog/gatra/scoring/strategy/GoldScoringStrategy.java
@@ -17,7 +17,7 @@ public class GoldScoringStrategy implements TierScoringStrategy {
     }
 
     @Override
-    public double calculateBaseScore(double totalPointsEarned, int totalMembers) {
+    public double calculateBaseScore(double totalPointsEarned, long totalMembers) {
         if (totalMembers <= 0) {
             return 0.0;
         }
